Update package versions in environment and requirements files
This commit updates the versions of several dependencies across the environment.yml, requirements_for_macOS.txt, and requirements.txt files. The Bio and biopython packages are upgraded to version 1.8.0 and 1.85 respectively, while biotite is updated to version 1.4.0. Additionally, a change in the data_utils.py file replaces the import of filter_backbone with filter_peptide_backbone to reflect the updated API.