Update package versions in requirements files for macOS and general environment
This commit upgrades several dependencies in both requirements_for_macOS.txt and requirements.txt. Notable updates include accelerate to 1.10.1, bitsandbytes to 0.47.0, datasets to 4.0.0, gradio to 5.44.1, matplotlib to 3.10.6, pandas to 2.3.2, peft to 0.17.1, scipy to 1.16.1, torchmetrics to 1.8.1, transformers to 4.48.1, and wandb to 0.21.3. Additionally, a new dependency, dotenv, has been added to both files.